Add support for finding MDIO bus interface by NIC name
Currently, aq-firmware-tool depends on mdio-proxy.ko, which finds
the MDIO bus interface through the MDIO bus name. But the current
mdio-proxy code is not in the Linux kernel, it is out-of-tree, so
we need to compile mdio-proxy.ko separately.

For some AQR PHYs, this method is necessary, because these PHYs
cannot be probed if they lack of firmware. But for some AQR PHYs,
like AQR 113C, it can be probed even if it without firmware. In
addtion, sometimes we may need to upgrade the firmware, so the
previous method is more cumbersome. For these situations, we can
use the existing dev_ioctl interface in the kernel to simplify
the process and find the MDIO bus interface through the NIC name,
so that mdio-proxy.ko is no longer needed.

Signed-off-by: Wei Fang <wei.fang@nxp.com>